requirements.txt
# Python ~= 3.6
attrs >= 18.1
click ~= 7.0
